llvm-project-rustc: LLVM-14.0 rust-specific Toolchain This is the rust enhanced/specific LLVM 14.0.0 toolchain, used in building/linking rust-lang projects.
We cannot use the existing OpenWrt eBPF LLVM toolchain for this.
The eBPF LLVM toolchain only builds for eBPF, but we need the
rust enhanced LLVM and support for AArch64, ARM, Mips, PowerPC,
and X86.

This LLVM is stored/preserved in staging_dir/host/llvm-rust and
carries over a clean, requiring only a build once situation
(unless the LLVM changes, or staging_dir gets removed).

Signed-off-by: Donald Hoskins <grommish@gmail.com>